requirements.txt
d2to1>=0.2.10,<0.3
pbr>=0.5
httplib2>=0.7.0
testtools>=0.9.32
testresources
testrepository
oslo.config>=1.1.0
numpy
matplotlib
numexpr>2.0.0
Cython>0.13
tables
